Removed version numbers from dependencies and placed tests in src/test/java
