+++ pom.xml
@@ -29,13 +29,11 @@
       <dependency>
           <groupId>org.springframework</groupId>
           <artifactId>spring-test</artifactId>
-          <version>5.0.8.RELEASE</version>
           <scope>test</scope>
       </dependency>
       <dependency>
           <groupId>org.springframework.boot</groupId>
           <artifactId>spring-boot-test-autoconfigure</artifactId>
-          <version>2.0.4.RELEASE</version>
           <scope>test</scope>
       </dependency>
   </dependencies>
