+++ src/test/java/org/broadleafcommerce/frameworkmapping/DisabledFrameworkApplication.java
@@ -1,4 +1,4 @@
-package org.broadleafcommerce;
+package org.broadleafcommerce.frameworkmapping;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
