+++ src/test/java/org/broadleafcommerce/frameworkmapping/DisabledFrameworkMappingTest.java
@@ -1,9 +1,8 @@
-package org.broadleafcommerce;
+package org.broadleafcommerce.frameworkmapping;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import org.broadleafcommerce.frameworkmapping.FrameworkControllerHandlerMapping;
 import org.broadleafcommerce.frameworkmapping.annotation.FrameworkController;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -15,27 +14,25 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 
-
 @WebMvcTest(includeFilters = {@ComponentScan.Filter(classes = FrameworkController.class),
                               @ComponentScan.Filter(type = FilterType.ASSIGNABLE_TYPE,
                                                     classes = FrameworkControllerHandlerMapping.class)})
-@ContextConfiguration(classes = EnabledFrameworkApplication.class)
+@ContextConfiguration(classes = DisabledFrameworkApplication.class)
 @RunWith(SpringRunner.class)
-public class EnabledFrameworkMappingTest {
+public class DisabledFrameworkMappingTest {
 
     @Autowired
     MockMvc mockMvc;
 
 
     @Test
-    public void testFrameworkOnlyGetMappingWorks() throws Exception {
+    public void testFrameworkOnlyGetMappingFails() throws Exception {
         mockMvc.perform(get("/framework-only-get"))
-                .andExpect(status().isOk())
-                .andExpect(content().string("frameworkControllerOnlyGetResponse"));
+                .andExpect(status().isNotFound());
     }
 
     @Test
-    public void testOverrideGetMappingWorks() throws Exception {
+    public void testOverrideGetMappingStillWorks() throws Exception {
         mockMvc.perform(get("/overridden-get"))
                 .andExpect(status().isOk())
                 .andExpect(content().string("customControllerGetResponse"));
