+++ src/test/java/org/broadleafcommerce/frameworkmapping/SubclassedCustomController.java
@@ -1,4 +1,4 @@
-package org.broadleafcommerce;
+package org.broadleafcommerce.frameworkmapping;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
